THRIFT-4645: TCurlClient: include failure reason in exception
When curl_exec fails, more detailed failure information is available
by calling curl_error.  Include this error information in the message in
the thrown TTransportException.

Also change the comparison of the return value of curl_exec to
explicitly check for boolean false, so as to distinguish from an empty
response body (per the PHP documentation on this subject).